+++ src/wallet/isWalletACoinbaseSmartWallet.ts
@@ -1,39 +1,58 @@
-import { decodeAbiParameters } from 'viem';
+import { checksumAddress, decodeAbiParameters } from 'viem';
 import {
+  CB_SW_FACTORY_ADDRESS,
   CB_SW_PROXY_BYTECODE,
   CB_SW_V1_IMPLEMENTATION_ADDRESS,
   ERC_1967_PROXY_IMPLEMENTATION_SLOT,
 } from './constants';
 import type { Address, BlockTag, Hex } from 'viem';
 import type {
-  IsWalletASmartWalletOptions,
-  IsWalletASmartWalletResponse,
+  IsWalletACoinbaseSmartWalletOptions,
+  IsWalletACoinbaseSmartWalletResponse,
 } from './types';
 
 /**
  * Validates a User Operation by checking if the sender address
  * is a proxy with the expected bytecode.
  */
-export async function isWalletASmartWallet({
+export async function isWalletACoinbaseSmartWallet({
   client,
   userOp,
-}: IsWalletASmartWalletOptions): Promise<IsWalletASmartWalletResponse> {
+}: IsWalletACoinbaseSmartWalletOptions): Promise<IsWalletACoinbaseSmartWalletResponse> {
   try {
     const code = await client.getBytecode({ address: userOp.sender });
+
+    if (!code) {
+      // no code at address, check that the initCode is deploying a Coinbase Smart Wallet
+      // factory address is first 20 bytes of initCode after '0x'
+      const factoryAddress = userOp.initCode.slice(0, 42) as Address;
+      if (
+        checksumAddress(factoryAddress) !==
+        checksumAddress(CB_SW_FACTORY_ADDRESS)
+      ) {
+        return {
+          isCoinbaseSmartWallet: false,
+          error: 'Invalid factory address',
+          code: 'W_ERR_1',
+        };
+      }
+      return { isCoinbaseSmartWallet: true };
+    }
+
     // Verify if the sender address bytecode matches the Coinbase Smart Wallet proxy bytecode
     if (code !== CB_SW_PROXY_BYTECODE) {
       return {
-        isSmartWallet: false,
+        isCoinbaseSmartWallet: false,
         error: 'Invalid bytecode',
-        code: 'W_ERR_1',
+        code: 'W_ERR_2',
       };
     }
   } catch (error) {
     console.error('Error retrieving bytecode:', error);
     return {
-      isSmartWallet: false,
+      isCoinbaseSmartWallet: false,
       error: 'Error retrieving bytecode',
-      code: 'W_ERR_2',
+      code: 'W_ERR_3',
     };
   }
 
@@ -49,9 +68,9 @@ export async function isWalletASmartWallet({
   } catch (error) {
     console.error('Error retrieving implementation address:', error);
     return {
-      isSmartWallet: false,
+      isCoinbaseSmartWallet: false,
       error: 'Error retrieving implementation address',
-      code: 'W_ERR_3',
+      code: 'W_ERR_4',
     };
   }
 
@@ -62,13 +81,16 @@ export async function isWalletASmartWallet({
   )[0];
 
   // Verify if the implementation address matches the expected Coinbase Smart Wallet address
-  if (implementationAddress !== CB_SW_V1_IMPLEMENTATION_ADDRESS) {
+  if (
+    checksumAddress(implementationAddress) !==
+    checksumAddress(CB_SW_V1_IMPLEMENTATION_ADDRESS)
+  ) {
     return {
-      isSmartWallet: false,
+      isCoinbaseSmartWallet: false,
       error: 'Invalid implementation address',
-      code: 'W_ERR_4',
+      code: 'W_ERR_5',
     };
   }
 
-  return { isSmartWallet: true };
+  return { isCoinbaseSmartWallet: true };
 }